feat: swagger spi sync
