+++ app/controller/sync.js
@@ -1,7 +1,6 @@
 'use strict';
 
-const schema = require('../common/schema');
-const { APISchema } = schema;
+const swagger = require('../service/swagger');
 
 /**
  * 同步 SPI Schema
@@ -11,59 +10,16 @@ const { APISchema } = schema;
  * this.request.body 获取请求body
  */
 exports.spiSync = async function (ctx) {
-  const API = ctx.model.Api;
-  const Project = ctx.model.Project;
-  const { appName, apis } = this.request.body;
-
-  const project = await Project.findOne({
-    name: appName,
-  });
-
-  // 1. 判断是否存在该后端应用
-  if (!project) {
+  try {
+    const result = await swagger.sync(ctx.model, ctx.request.body);
+    this.body = {
+      status: 'success',
+      data: result,
+    };
+  } catch (err) {
     this.body = {
       status: 'fail',
-      data: {},
-      msg: '该应用不存在',
+      msg: err.message,
     };
-    return;
   }
-
-  const actions = [];
-
-  // const apiList = JSON.parse(apis);
-  const apiList = apis;
-  const projectApis = await API.find({ projectId: project._id });
-  const projectApiUrls = projectApis.map(item => item.url);
-
-  apiList.forEach(item => {
-    let action = null;
-    // 判断接口是否存在，存在走更新，不存在走创建
-    if (!item.bizType) return;
-
-    if (projectApiUrls.includes(item.bizType)) {
-      action = API.updateOne({ url: item.bizType }, { $set: {
-        requestAutoSchema: item.requestSchema,
-        responseAutoSchema: item.responseSchema,
-      } });
-    } else {
-      action = API.create({
-        ...APISchema,
-        projectId: project._id.toString(),
-        apiType: 'SPI',
-        url: item.bizType,
-        requestAutoSchema: item.requestSchema,
-        responseAutoSchema: item.responseSchema,
-      });
-    }
-
-    actions.push(action);
-  });
-
-  const result = await Promise.all(actions);
-
-  this.body = {
-    status: 'success',
-    data: result,
-  };
 };
